+++ src/umi-isolation.py
@@ -22,14 +22,13 @@
 
 # Imports:
 import os
-import sys
 import argparse
 import re
 import pandas as pd
 import subprocess as sp
 
 
-def create_output_files(cluster_directory, blast_file, tabular_file):
+def create_output_files(cluster_directory, output_blast_file, tabular_file):
     """
     The create_output_files function:
         This function creates the tabular and BLAST output files.
@@ -57,7 +56,7 @@ def create_output_files(cluster_directory, blast_file, tabular_file):
                             header.strip("\n"),
                             read.strip("\n").upper(),
                         ]
-                        with open(blast_file, "a") as output_file:
+                        with open(output_blast_file, "a") as output_file:
                             output_file.write(">" + umi_number + "\n")
                             output_file.write(read.strip("\n").upper() + "\n")
                     else:
@@ -75,7 +74,7 @@ def create_output_files(cluster_directory, blast_file, tabular_file):
                             header.strip("\n"),
                             read.strip("\n").upper(),
                         ]
-                        with open(blast_file, "a") as output_file:
+                        with open(output_blast_file, "a") as output_file:
                             output_file.write(">" + umi_version + "\n")
                             output_file.write(read.strip("\n").upper() + "\n")
                         version_count += 1
@@ -89,7 +88,7 @@ def create_output_files(cluster_directory, blast_file, tabular_file):
     output.to_csv(tabular_file, sep="\t", encoding="utf-8")
 
 
-def get_vsearch_cluster_size(zip, cluster_directory, identity_score):
+def get_vsearch_cluster_size(zip_file, cluster_directory, identity_score):
     """
     The get_vsearch_cluster_size function:
         This function controls the VSEARCH clustering. Every fasta file created
@@ -97,9 +96,9 @@ def get_vsearch_cluster_size(zip, cluster_directory, identity_score):
         result is a single centroid sequence. This is checked in the
         create_output_files function.
     """
-    for file_name in os.listdir(zip):
+    for file_name in os.listdir(zip_file):
         if file_name.startswith("sorted"):
-            input_command = zip + file_name
+            input_command = zip_file + file_name
             output_command = cluster_directory + file_name[11:]
             vsearch_cluster_command = sp.Popen(
                 [
@@ -125,17 +124,17 @@ def get_vsearch_cluster_size(zip, cluster_directory, identity_score):
             pass
 
 
-def get_vsearch_sort_by_size(zip, minimal_size_abundance):
+def get_vsearch_sort_by_size(zip_file, minimal_size_abundance):
     """
     The get_vsearch_sort_by_size function:
         This function controls the VSEARCH sorting. Every fasta file created by
         get_vsearch_derep is sorted based on abundance. Any reads with a
         abundance lower than minimal_size_abundance will be discarded.
     """
-    for file_name in os.listdir(zip):
+    for file_name in os.listdir(zip_file):
         if file_name.startswith("derep"):
-            input_command = zip + file_name
-            output_command = zip + "sorted" + file_name
+            input_command = zip_file + file_name
+            output_command = zip_file + "sorted" + file_name
             vsearch_sort_command = sp.Popen(
                 [
                     "vsearch",
@@ -156,17 +155,17 @@ def get_vsearch_sort_by_size(zip, minimal_size_abundance):
             pass
 
 
-def get_vsearch_derep(zip):
+def get_vsearch_derep(zip_file):
     """
     The get_vsearch_derep function:
         This function controls the VSEARCH dereplication. Every fasta file
         created by get_fasta_files is dereplicated. This step is necessary for
         the sorting step to work.
     """
-    for file_name in os.listdir(zip):
+    for file_name in os.listdir(zip_file):
         if file_name.endswith(".fasta"):
-            input_command = zip + file_name
-            output_command = zip + "derep" + file_name
+            input_command = zip_file + file_name
+            output_command = zip_file + "derep" + file_name
             vsearch_derep_command = sp.Popen(
                 [
                     "vsearch",
@@ -187,7 +186,7 @@ def get_vsearch_derep(zip):
             pass
 
 
-def get_fasta_files(header, read, umi_code, unique_umi_dictionary, zip):
+def get_fasta_files(header, read, umi_code, unique_umi_dictionary, zip_file):
     """
     The get_fasta_files function:
         This function creates separate fasta files for every unique UMI. The
@@ -202,7 +201,7 @@ def get_fasta_files(header, read, umi_code, unique_umi_dictionary, zip):
         + umi_code
         + ".fasta"
     )
-    file_name = zip + file_identifier
+    file_name = zip_file + file_identifier
     with open(file_name, "a") as output_file:
         output_file.write(header)
         output_file.write(read)
@@ -236,96 +235,103 @@ def get_target_zero(read, umi_length, search_method, forward, reverse):
         pass
 
 
-# The getTargetFront function.
-# This function searches for a regex string in the provided read. It will
-# isolate either a forward or reverse UMI or double UMIs. The isolation is based on
-# this read structure SCAFFOLDF-UMI-PRIMERF-PRODUCT-PRIMERR-UMI-SCAFFOLDR.
-# When looking for the forward UMI, the last position of SCAFFOLDF is used, when
-# looking for the reverse UMI, the first position of SCAFFOLDR is used, when
-# looking for double UMIs both positions are used.
-# The mentioned positions + or - the UMI length result in a UMI code.
-# When not searching for both UMIs a check needs to be passed, this check
-# makes sure the reverse scaffold (in the case of umi5) or the forward scaffold
-# (in the case of umi3) are present.
-def getTargetFront(read, umi_length, search_method, forward, reverse):
+def get_target_front(read, umi_length, search_method, forward, reverse):
+    """
+    The get_target_front function:
+        This function searches for a regex string in the provided read. It will
+        isolate either a forward or reverse UMI or double UMIs. The isolation
+        is based on this read structure
+        SCAFFOLDF-UMI-PRIMERF-PRODUCT-PRIMERR-UMI-SCAFFOLDR. When looking for
+        the forward UMI, the last position of SCAFFOLDF is used, when looking
+        for the reverse UMI, the first position of SCAFFOLDR is used, when
+        looking for double UMIs both positions are used. The mentioned
+        positions + or - the UMI length result in a UMI code. When not
+        searching for both UMIs a check needs to be passed, this check makes
+        sure the reverse scaffold (in the case of umi5) or the forward
+        scaffold (in the case of umi3) are present.
+    """
     if search_method == "umi5" or search_method == "umidouble":
-        intPositionForward = re.search(forward, read).end()
-        intPositionUmiForward = intPositionForward + int(umi_length)
-        strUmiCodeForward = read[intPositionForward:intPositionUmiForward]
+        forward_position = re.search(forward, read).end()
+        umi_forward_position = forward_position + int(umi_length)
+        forward_umi_code = read[forward_position:umi_forward_position]
         if search_method == "umi5":
             check_reverse_primer = re.search(reverse, read)
             if check_reverse_primer != None:
-                return strUmiCodeForward
+                return forward_umi_code
             else:
                 pass
         elif search_method == "umidouble":
-            intPositionReverse = re.search(reverse, read).start()
-            intPositionUmiReverse = intPositionReverse - int(umi_length)
-            strUmiCodeReverse = read[intPositionUmiReverse:intPositionReverse]
-            return strUmiCodeForward, strUmiCodeReverse
+            reverse_position = re.search(reverse, read).start()
+            umi_reverse_position = reverse_position - int(umi_length)
+            reverse_umi_code = read[umi_reverse_position:reverse_position]
+            return forward_umi_code, reverse_umi_code
         else:
             pass
     elif search_method == "umi3":
         check_forward_primer = re.search(forward, read)
         if check_forward_primer != None:
-            intPositionReverse = re.search(reverse, read).start()
-            intPositionUmiReverse = intPositionReverse - int(umi_length)
-            strUmiCodeReverse = read[intPositionUmiReverse:intPositionReverse]
-            return strUmiCodeReverse
+            reverse_position = re.search(reverse, read).start()
+            umi_reverse_position = reverse_position - int(umi_length)
+            reverse_umi_code = read[umi_reverse_position:reverse_position]
+            return reverse_umi_code
         else:
             pass
     else:
         pass
 
 
-# The getTargetBehind function.
-# This function searches for a regex string in the provided read. It will
-# isolate either a forward or reverse UMI or double UMIs. The isolation is based on
-# this read structure UMI-PRIMERF-PRODUCT-PRIMERR-UMI.
-# When looking for the forward UMI, the first position of PRIMERF is used, when
-# looking for the reverse UMI, the last position of PRIMERR is used, when
-# looking for double UMIs both positions are used.
-# The mentioned positions + or - the UMI length result in a UMI code.
-# When not searching for both UMIs a check needs to be passed, this check
-# makes sure the reverse primer (in the case of umi5) or the forward primer
-# (in the case of umi3) are present.
-def getTargetBehind(read, umi_length, search_method, forward, reverse):
+def get_target_behind(read, umi_length, search_method, forward, reverse):
+    """
+    The get_target_behind function:
+        This function searches for a regex string in the provided read. It will
+        isolate either a forward or reverse UMI or double UMIs. The isolation
+        is based on this read structure UMI-PRIMERF-PRODUCT-PRIMERR-UMI. When
+        looking for the forward UMI, the first position of PRIMERF is used,
+        when looking for the reverse UMI, the last position of PRIMERR is used,
+        when looking for double UMIs both positions are used. The mentioned
+        positions + or - the UMI length result in a UMI code. When not
+        searching for both UMIs a check needs to be passed, this check makes
+        sure the reverse primer (in the case of umi5) or the forward primer
+        (in the case of umi3) are present.
+    """
     if search_method == "umi5" or search_method == "umidouble":
-        intPositionForward = re.search(forward, read).start()
-        intPositionUmiForward = intPositionForward - int(umi_length)
-        strUmiCodeForward = read[intPositionUmiForward:intPositionForward]
+        forward_position = re.search(forward, read).start()
+        umi_forward_position = forward_position - int(umi_length)
+        forward_umi_code = read[umi_forward_position:forward_position]
         if search_method == "umi5":
             check_reverse_primer = re.search(reverse, read)
             if check_reverse_primer != None:
-                return strUmiCodeForward
+                return forward_umi_code
             else:
                 pass
         elif search_method == "umidouble":
-            intPositionReverse = re.search(reverse, read).end()
-            intPositionUmiReverse = intPositionReverse + int(umi_length)
-            strUmiCodeReverse = read[intPositionReverse:intPositionUmiReverse]
-            return strUmiCodeForward, strUmiCodeReverse
+            reverse_position = re.search(reverse, read).end()
+            umi_reverse_position = reverse_position + int(umi_length)
+            reverse_umi_code = read[reverse_position:umi_reverse_position]
+            return forward_umi_code, reverse_umi_code
         else:
             pass
     elif search_method == "umi3":
         check_forward_primer = re.search(forward, read)
         if check_forward_primer != None:
-            intPositionReverse = re.search(reverse, read).end()
-            intPositionUmiReverse = intPositionReverse + int(umi_length)
-            strUmiCodeReverse = read[intPositionReverse:intPositionUmiReverse]
-            return strUmiCodeReverse
+            reverse_position = re.search(reverse, read).end()
+            umi_reverse_position = reverse_position + int(umi_length)
+            reverse_umi_code = read[reverse_position:umi_reverse_position]
+            return reverse_umi_code
         else:
             pass
     else:
         pass
 
 
-# The getReverseComplement function.
-# This function creates a complementary string using a sequence as input. The
-# function loops through a list version of the sequence and checks and changes
-# every character. It then returns a joined string.
-def getReverseComplement(line):
-    dicComplementCodes = {
+def create_reverse_complement(line):
+    """
+    The create_reverse_complement function:
+        This function creates a complementary string using a sequence as input.
+        The function loops through a list version of the sequence and checks
+        and changes every character. It then returns a joined string.
+    """
+    complement_codes_dictionary = {
         "A": "T",
         "T": "A",
         "G": "C",
@@ -342,20 +348,21 @@ def getReverseComplement(line):
         "B": "V",
         "N": "N",
     }
-    lstLine = list(line)
-    for intPosition in range(len(lstLine)):
-        lstLine[intPosition] = dicComplementCodes[lstLine[intPosition]]
-    return "".join(lstLine)
+    line_list = list(line)
+    for position in range(len(line_list)):
+        line_list[position] = complement_codes_dictionary[line_list[position]]
+    return "".join(line_list)
 
-
-# The getRegex function.
-# This function creates a regex string using a sequence as input. This regex
-# string is based on the IUPAC ambiguity codes. The function loops through
-# a list version of the sequence and checks per character if it is a
-# ambiguous character, and changes it into regex code if so. It then returns
-# a joined string.
-def getRegex(line):
-    dicAmbiguityCodes = {
+def generate_regex(line):
+    """
+    The generate_regex function:
+        This function creates a regex string using a sequence as input. This
+        regex string is based on the IUPAC ambiguity codes. The function loops
+        through a list version of the sequence and checks per character if it
+        is a ambiguous character, and changes it into regex code if so. It then
+        returns a joined string.
+    """
+    ambiguity_codes = {
         "M": "[AC]",
         "R": "[AG]",
         "W": "[AT]",
@@ -368,123 +375,130 @@ def getRegex(line):
         "B": "[CGT]",
         "N": "[GATC]",
     }
-    lstLine = list(line)
-    for intPosition in range(len(lstLine)):
+    line_list = list(line)
+    for position in range(len(line_list)):
         if (
-            lstLine[intPosition] != "A"
-            and lstLine[intPosition] != "T"
-            and lstLine[intPosition] != "G"
-            and lstLine[intPosition] != "C"
+            line_list[position] != "A"
+            and line_list[position] != "T"
+            and line_list[position] != "G"
+            and line_list[position] != "C"
         ):
-            lstLine[intPosition] = dicAmbiguityCodes[lstLine[intPosition]]
+            line_list[position] = ambiguity_codes[line_list[position]]
         else:
             pass
-    return "".join(lstLine)
+    return "".join(line_list)
 
 
-# The getUmiCode function.
-# This function controls the umi searching process. It first uses the functions
-# getRegex and getReverseComplement to create regex strings of both the
-# forward primer/scaffold and the reverse complement primer/scaffold. These regex
-# strings are then directed to the desired functions, this depends on the
-# search method choice [primer/scaffold/zero].
-def getUmiCode(read, strProcess, umi_length, search_method, forward, reverse):
+def get_umi_code(read, process, umi_length, search_method, forward, reverse):
+    """
+    The get_umi_code function:
+        This function controls the umi searching process. It first uses the
+        functions generate_regex and create_reverse_complement to create regex
+        strings of both the forward primer/scaffold and the reverse complement
+        primer/scaffold. These regex strings are then directed to the desired
+        functions, this depends on the search method
+        choice [primer/scaffold/zero].
+    """
     read = read.strip("\n")
-    strRegexForward = getRegex(forward)
-    strRegexComplementReverse = getRegex(getReverseComplement(reverse[::-1]))
-    if strProcess == "primer":
+    forward_regex = generate_regex(forward)
+    reverse_complement_regex = generate_regex(
+        create_reverse_complement(reverse[::-1])
+    )
+    if process == "primer":
         try:
-            return getTargetBehind(
+            return get_target_behind(
                 read,
                 umi_length,
                 search_method,
-                strRegexForward,
-                strRegexComplementReverse,
+                forward_regex,
+                reverse_complement_regex,
             )
         except AttributeError:
             pass
-    elif strProcess == "scaffold":
+    elif process == "scaffold":
         try:
-            return getTargetFront(
+            return get_target_front(
                 read,
                 umi_length,
                 search_method,
-                strRegexForward,
-                strRegexComplementReverse,
+                forward_regex,
+                reverse_complement_regex,
             )
         except AttributeError:
             pass
-    elif strProcess == "zero":
+    elif process == "zero":
         try:
             return get_target_zero(
                 read,
                 umi_length,
                 search_method,
-                strRegexForward,
-                strRegexComplementReverse,
+                forward_regex,
+                reverse_complement_regex,
             )
         except AttributeError:
             pass
     else:
         pass
 
-
-# The getUmiCollection function.
-# This function opens the input file and loops through it. It isolates the
-# read headers and reads. For every read the getUmiCode function is called
-# which outputs one or two UMIs. In the case of a double UMI search [umidouble]
-# the two UMIs are put together. For every read that contains a UMI the
-# get_fasta_files function is called. After all reads have been processed, the
-# getVSEARCH and create_output_files functions are called.
-def getUmiCollection(
-    flInput,
+def get_umi_collection(
+    input_file,
     cluster_directory,
     tabular_file,
-    zip,
-    blast_file,
-    strProcess,
+    zip_file,
+    output_blast_file,
+    process,
     umi_length,
     search_method,
     forward,
     reverse,
-    strFormat,
-    strOperand,
+    format_string,
+    operand,
     identity_score,
     minimal_size_abundance,
 ):
+    """
+    The get_umi_collection function:
+        This function opens the input file and loops through it. It isolates the
+        read headers and reads. For every read the get_umi_code function is
+        called which outputs one or two UMIs. In the case of a double UMI
+        search [umidouble] the two UMIs are put together. For every read that
+        contains a UMI the get_fasta_files function is called. After all reads
+        have been processed, the VSEARCH and create_output_files functions
+        are called.
+    """
     unique_umi_dictionary = {}
-    intNoUmiInReadCount = 0
-    intUniqueUmi = 1
-    with open(flInput) as oisInput:
-        for line in oisInput:
+    count_reads_without_umi = 0
+    count_unique_umis = 1
+    with open(input_file) as input:
+        for line in input:
             if (
-                line[0] == strOperand
+                line[0] == operand
                 and bool(re.match("[A-Za-z0-9]", line[1])) == True
             ):
                 header = line
-                read = next(oisInput)
+                read = next(input)
                 try:
-                    strUmiCode = getUmiCode(
+                    umi_code = get_umi_code(
                         read.upper(),
-                        strProcess,
+                        process,
                         umi_length,
                         search_method,
                         forward.upper(),
                         reverse.upper(),
                     )
                 except UnboundLocalError:
-                    intNoUmiInReadCount += 1
+                    count_reads_without_umi += 1
                 try:
-                    if strUmiCode != None:
+                    if umi_code != None:
                         if search_method == "umi5" or search_method == "umi3":
-                            umi_code = strUmiCode
+                            umi_code = umi_code
                         elif search_method == "umidouble":
-                            umi_code = strUmiCode[0] + strUmiCode[1]
+                            umi_code = umi_code[0] + umi_code[1]
                         else:
                             pass
                         if umi_code not in unique_umi_dictionary:
-                            unique_umi_dictionary[umi_code] = intUniqueUmi
-                            intUniqueUmi += 1
+                            unique_umi_dictionary[umi_code] = count_unique_umis
+                            count_unique_umis += 1
                         else:
                             pass
                     else:
@@ -494,31 +508,32 @@ def getUmiCollection(
                 try:
                     if umi_code != None:
                         get_fasta_files(
-                            header, read, umi_code, unique_umi_dictionary, zip
+                            header, read, umi_code, unique_umi_dictionary, zip_file
                         )
                     else:
                         pass
                 except UnboundLocalError:
                     pass
-            strUmiCode = None
             umi_code = None
-    get_vsearch_derep(zip)
-    get_vsearch_sort_by_size(zip, minimal_size_abundance)
-    get_vsearch_cluster_size(zip, cluster_directory, identity_score)
-    create_output_files(cluster_directory, blast_file, tabular_file)
+            umi_code = None
+    get_vsearch_derep(zip_file)
+    get_vsearch_sort_by_size(zip_file, minimal_size_abundance)
+    get_vsearch_cluster_size(zip_file, cluster_directory, identity_score)
+    create_output_files(cluster_directory, output_blast_file, tabular_file)
 
-
-# The setFormat function.
-# This function specifies the first character of the read headers based on the
-# input file format. It then calls the getUmiCollection function.
-def setFormat(
-    flInput,
+def set_format(
+    """
+    The set_format function:
+        This function specifies the first character of the read headers based
+        on the input file format. It then calls the get_umi_collection function.
+    """
+    input_file,
     cluster_directory,
     tabular_file,
-    zip,
-    blast_file,
-    strProcess,
-    strFormat,
+    zip_file,
+    output_blast_file,
+    process,
+    format_string,
     umi_length,
     search_method,
     forward,
@@ -526,119 +541,121 @@ def setFormat(
     identity_score,
     minimal_size_abundance,
 ):
-    if strFormat == "fasta":
-        strOperand = ">"
-    elif strFormat == "fastq":
-        strOperand = "@"
+    if format_string == "fasta":
+        operand = ">"
+    elif format_string == "fastq":
+        operand = "@"
     else:
         pass
-    getUmiCollection(
-        flInput,
+    get_umi_collection(
+        input_file,
         cluster_directory,
         tabular_file,
-        zip,
-        blast_file,
-        strProcess,
+        zip_file,
+        output_blast_file,
+        process,
         umi_length,
         search_method,
         forward,
         reverse,
-        strFormat,
-        strOperand,
+        format_string,
+        operand,
         str(identity_score),
         str(minimal_size_abundance),
     )
 
 
-# The argvs function.
-def parseArgvs():
+def parse_argvs():
+    """
+    The parse_argvs function:
+        This function handles all positional arguments that the script accepts,
+        including version and help pages.
+    """
     parser = argparse.ArgumentParser(
-        description="Use a python script to\
-                                                  accumulate all UMIs and\
-                                                  output a tabular file, a\
-                                                  BLAST file and a zip file."
+        description="Use a python script to accumulate all UMIs and output a\
+                     tabular file, a BLAST file and a zip file."
     )
     parser.add_argument("-v", action="version", version="%(prog)s [0.1.0]")
     parser.add_argument(
         "-i",
         action="store",
-        dest="fisInput",
+        dest="input_file",
         help="The location of the input file(s)",
     )
     parser.add_argument(
         "-c",
         action="store",
-        dest="fosClusterDirectory",
+        dest="cluster_directory",
         help="The location of the clustering output file(s)",
     )
     parser.add_argument(
         "-o",
         action="store",
-        dest="fosOutput",
+        dest="output_tabular_file",
         help="The location of the tabular output file(s)",
     )
     parser.add_argument(
         "-z",
         action="store",
-        dest="fosOutputZip",
+        dest="output_zip_file",
         help="The location of the pre-vsearch zip output file(s)",
     )
     parser.add_argument(
         "-q",
         action="store",
-        dest="fosBlastFile",
+        dest="output_blast_file",
         help="The location of the BLAST output file(s)",
     )
     parser.add_argument(
         "-p",
         action="store",
-        dest="disProcess",
+        dest="process",
         help="The UMI search approach [primer/scaffold(adapter)/zero]",
     )
     parser.add_argument(
         "-f",
         action="store",
-        dest="disFormat",
+        dest="format",
         help="The format of the input file(s) [fasta/fastq]",
     )
     parser.add_argument(
         "-l",
         action="store",
-        dest="disUmiLength",
+        dest="umi_length",
         help="The length of the UMI sequences",
     )
     parser.add_argument(
         "-s",
         action="store",
-        dest="disSearch",
-        help="Search UMIs at 5'-end [umi5], 3'-end [umi3] or at\
-                              5'-end and 3'-end [umidouble]",
+        dest="search_method",
+        help="Search UMIs at 5'-end [umi5], 3'-end [umi3] or at 5'-end and\
+              3'-end [umidouble]",
     )
     parser.add_argument(
         "-a",
         action="store",
-        dest="disForward",
+        dest="forward",
         help="The 5'-end search nucleotides",
     )
     parser.add_argument(
         "-b",
         action="store",
-        dest="disReverse",
+        dest="reverse",
         help="The 3'-end search nucleotides",
     )
     parser.add_argument(
         "-d",
         action="store",
-        dest="disIdentity",
-        help="The identity percentage with which to perform the\
-                              final VSEARCH check",
+        dest="identity_score",
+        help="The identity percentage with which to perform the final VSEARCH\
+              check",
     )
     parser.add_argument(
         "-u",
         action="store",
-        dest="disAbundance",
-        help="The minimum abundance a read has to be present in\
-                              order to be part of the final VSEARCH check",
+        dest="abundance",
+        help="The minimum abundance a read has to be present in order to be\
+              part of the final VSEARCH check",
     )
     argvs = parser.parse_args()
     return argvs
@@ -648,23 +665,23 @@ def main():
     """
     The main function:
         This function handles the arguments parsed to the script and calls
-        the first function setFormat.
+        the first function set_format.
     """
     argvs = parseArgvs()
-    setFormat(
-        argvs.fisInput,
-        argvs.fosClusterDirectory,
-        argvs.fosOutput,
-        argvs.fosOutputZip,
-        argvs.fosBlastFile,
-        argvs.disProcess,
-        argvs.disFormat,
-        argvs.disUmiLength,
-        argvs.disSearch,
-        argvs.disForward,
-        argvs.disReverse,
-        argvs.disIdentity,
-        argvs.disAbundance,
+    set_format(
+        argvs.input_file,
+        argvs.cluster_directory,
+        argvs.output_tabular_file,
+        argvs.output_zip_file,
+        argvs.output_blast_file,
+        argvs.process,
+        argvs.format,
+        argvs.umi_length,
+        argvs.search_method,
+        argvs.forward,
+        argvs.reverse,
+        argvs.identity_score,
+        argvs.abundance,
     )
 
 
@@ -674,11 +691,11 @@ if __name__ == "__main__":
 # Additional information:
 # =======================
 # Prequisites:
-#     sudo apt-get install python3
-#     sudo apt-get install python3-pip
-#     sudo pip3 install pandas
-#     sudo apt-get install libargtable2-dev
-#     Download VSEARCH from GitHub
-#     Unpack downloaded file
-#     sudo ./autogen.sh
-#     sudo ./configure && sudo make && sudo make install
+# * sudo apt-get install python3
+# * sudo apt-get install python3-pip
+# * sudo pip3 install pandas
+# * sudo apt-get install libargtable2-dev
+# * Download VSEARCH from GitHub
+# * Unpack downloaded file
+# * sudo ./autogen.sh
+# * sudo ./configure && sudo make && sudo make install